+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 Notable changes to this project.
 
-## [dev] - unreleased
+## [1.5.5] - 2019-06-13
 - include `reputationPayments` in `get_payments`
 - add datetime information (`insertedAt`) to `get_transactions`